requirements.txt
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
python-dotenv==0.20.0
python-telegram-bot==20.0a0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
